# apidocs/swagger.yaml
basePath: /
definitions:
  github_com_stockflow_server_internal_response.Envelope:
    properties:
      code:
        description: 成功 0（数字）；失败为模块命名空间错误码字符串（COMMON_* 等）
      data: {}
      details:
        description: 校验失败必带（api.md §4）
      message:
        type: string
      request_id:
        type: string
    type: object
...
  internal_returns.CancelInput:
    properties:
      reason:
        type: string
    type: object
  internal_returns.ExceptionAssignInput:
    properties:
      assignee_id:
        type: integer
      assignee_name:
        type: string
    type: object
  internal_returns.ExceptionCreateInput:
    properties:
      bin_id:
        description: 可空定位：0=未定位
        type: integer
      detail:
        type: string
      freeze_batch_id:
        description: 0=非批次 SKU
        type: integer
      freeze_enabled:
        description: |-
          冻结选项：定位到具体库存行时才可冻结（仓库+库位+SKU 定位，inventory-rules §4.1）；
          FreezeQty 为冻结量（解冻释放取量依据）。
        type: boolean
      freeze_qty:
        type: string
      freeze_warehouse_id:
        type: integer
      owner_id:
        type: integer
      owner_name:
        type: string
      remark:
        type: string
      serial_no:
        type: string
      sku_id:
        description: 可空定位：0=未定位
        type: integer
      source_no:
        type: string
      source_type:
        description: 产生环节的单据类型（收货/拣货/复核/盘点等）
        type: string
      type:
        description: 九类中文值域（business-flow §11.2）
        type: string
    type: object
  internal_returns.PurchaseReturnCreateInput:
    properties:
      lines:
        items:
          $ref: '#/definitions/internal_returns.PurchaseReturnLineInput'
        type: array
      po_no:
        type: string
      remark:
        type: string
      supplier_id:
        type: integer
      warehouse_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: StockFlow（库流智能仓储管理系统）后端 API。统一信封 {code,message,data,request_id,details}（api.md
    §2）。
  title: StockFlow API
  version: "1.0"
paths:
...
  /api/exceptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/exceptions
      tags:
      - 退货与异常
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_returns.ExceptionCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/exceptions
      tags:
      - 退货与异常
  /api/exceptions/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/exceptions/:id
      tags:
      - 退货与异常
  /api/exceptions/{id}/assign:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_returns.ExceptionAssignInput'
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/exceptions/:id/assign
      tags:
      - 退货与异常
  /api/exceptions/{id}/close:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/exceptions/:id/close
      tags:
      - 退货与异常
  /api/exceptions/{id}/resolve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/exceptions/:id/resolve
      tags:
      - 退货与异常
  /api/exceptions/{id}/review:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/exceptions/:id/review
      tags:
      - 退货与异常
  /api/exceptions/{id}/start:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/exceptions/:id/start
      tags:
      - 退货与异常
...
  /api/inventory/trace:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/inventory/trace
      tags:
      - 退货与异常
...
  /api/purchase-returns:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/purchase-returns
      tags:
      - 退货与异常
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_returns.PurchaseReturnCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchase-returns
      tags:
      - 退货与异常
  /api/purchase-returns/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/purchase-returns/:id
      tags:
      - 退货与异常
  /api/purchase-returns/{id}/approve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_returns.ApproveInput'
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchase-returns/:id/approve
      tags:
      - 退货与异常
  /api/purchase-returns/{id}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_returns.CancelInput'
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchase-returns/:id/cancel
      tags:
      - 退货与异常
  /api/purchase-returns/{id}/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchase-returns/:id/complete
      tags:
      - 退货与异常
  /api/purchase-returns/{id}/ship:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_returns.ShipInput'
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchase-returns/:id/ship
      tags:
      - 退货与异常
  /api/purchase-returns/{id}/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: POST /api/purchase-returns/:id/submit
      tags:
      - 退货与异常
...
  /api/returns/{id}:
    get:
      parameters:
      - description: 路径参数 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 统一响应信封
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_stockflow_server_internal_response.Envelope'
      summary: GET /api/purchase-returns/:id
      tags:
      - 退货与异常
...
securityDefinitions:
  BearerAuth:
    description: 值格式 "Bearer <access_token>"（api.md §6.1）。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"